rtdls_project/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
django-filter==24.3
dj-database-url==2.3.0
channels==4.1.0
channels-redis==4.2.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
gunicorn==23.0.0
uvicorn==0.34.0
reportlab==4.2.5
openpyxl==3.1.5
drf-spectacular==0.28.0
